pyForms/requirements.txt
Django
Pillow
django-crispy-forms
crispy-bootstrap5
python-decouple
